+++ src/tv_show/episodes.rs
@@ -4,19 +4,19 @@ use std::sync::Arc;
 
 use tracing::instrument;
 
-use crate::api_structs;
+use super::api_structs;
 use crate::error::{Error, Result};
 use crate::http_client::HttpClientTrait;
-use crate::models::Episode;
+use crate::models::MediaItem;
 
 const TV3_EPISODE_LIST_URL: &str = "https://www.3cat.cat/api/3cat/dades/?queryKey=%5B%22tira%22%2C%7B%22url%22%3A%22%2F%2Fapi.3cat.cat%2Fvideos%3F_format%3Djson%26no_agrupacio%3DPUAGR_LLSIGN%26tipus_contingut%3DPPD%26items_pagina%3D1500%26pagina%3D1%26sdom%3Dimg%26version%3D2.0%26cache%3D180%26https%3Dtrue%26master%3Dyes%26programatv_id%3D{tv_show_id}%26origen%3Dauto%26perfil%3Dpc%22%7D%5D";
 
 #[instrument(skip(http_client))]
-pub(crate) async fn get_episodes<T>(http_client: &Arc<T>, tv_show_id: i32) -> Result<Vec<Episode>>
+pub(crate) async fn get_episodes<T>(http_client: &Arc<T>, tv_show_id: i32) -> Result<Vec<MediaItem>>
 where
     T: HttpClientTrait,
 {
-    let mut episodes: Vec<Episode> = vec![];
+    let mut episodes: Vec<MediaItem> = vec![];
 
     let url = TV3_EPISODE_LIST_URL.replace("{tv_show_id}", &tv_show_id.to_string());
 
@@ -36,13 +36,13 @@ where
             _ => item.permatitle,
         };
 
-        episodes.push(Episode {
+        episodes.push(MediaItem {
             id: item.id,
             title,
             video_url: None,
             subtitle_url: None,
-            episode_number: item.number_of_episode,
-            tv_show_name: item.tv_show_name,
+            episode_number: Some(item.number_of_episode),
+            tv_show_name: Some(item.tv_show_name),
         });
     }